closing feature #7035 QuickInfo : managing attached files
Feature #7035: attachments on news.
Wiring the attachments to the id of the publication linked to the news instead of the id of the news.
Simplifying some code.
Fixing a little UI problem, on a news from the list of the homepage of quickinfo application, when comments are not enabled and no attachment exist.
Fixing the portlet in order to not display comment icon and comment number when comments are not enabled from the application parameters.
Replacing java scriptlet by JSTL instructions.

closing #7035 fixing pb of thumbnail display